discord-mod-manager/requirements.txt
discord.py==2.4.0
requests
aiosqlite
python-dotenv